Renames after pull request comments.
